third_party/blink/web_tests/external/wpt/upgrade-insecure-requests/gen/worker-module.http-rp/upgrade/worker-classic.https-expected.txt

This is a testharness.js-based test.
[FAIL] Upgrade-Insecure-Requests: Expects allowed for worker-classic to same-http-downgrade origin and downgrade redirection from https context.
  assert_equals: The resource request should be 'allowed'. expected "allowed" but got "blocked"
[FAIL] Upgrade-Insecure-Requests: Expects allowed for worker-classic to same-http-downgrade origin and no-redirect redirection from https context.
  assert_equals: The resource request should be 'allowed'. expected "allowed" but got "blocked"
Harness: the test ran to completion.
